@@ -19,9 +19,10 @@ disallow_untyped_defs = "True"
 "Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22langchain-mongodb%3D%3D0%22&expanded=true"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<4.0"
 pymongo = ">=4.6.1,<5.0"
 langchain-core = "^0.2.21"
+pydantic = ">=2,<3"
 [[tool.poetry.dependencies.numpy]]
 version = "^1"
 python = "<3.12"
