@@ -12,10 +12,11 @@ license = "MIT"
 "Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22langchain-prompty%3D%3D0%22&expanded=true"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<4.0"
 langchain-core = "^0.2.36"
 pyyaml = "^6.0.1"
 types-pyyaml = "^6.0.12.20240311"
+pydantic = ">=2,<3"
 
 [tool.poetry.group.test]
 optional = true
