@@ -1,9 +1,10 @@
-from typing import Any, Dict, List, Optional
+from typing import List, Optional
 
 from langchain_core.callbacks import CallbackManagerForRetrieverRun
 from langchain_core.documents import Document
-from langchain_core.pydantic_v1 import root_validator
 from langchain_core.retrievers import BaseRetriever
+from pydantic import ConfigDict, model_validator
+from typing_extensions import Self
 
 from langchain_box.utilities import BoxAuth, _BoxAPIWrapper
 
@@ -129,30 +130,31 @@ class BoxRetriever(BaseRetriever):
 
     _box: Optional[_BoxAPIWrapper]
 
-    class Config:
-        arbitrary_types_allowed = True
-        extra = "allow"
+    model_config = ConfigDict(
+        arbitrary_types_allowed=True,
+        extra="allow",
+    )
 
-    @root_validator(allow_reuse=True)
-    def validate_box_loader_inputs(cls, values: Dict[str, Any]) -> Dict[str, Any]:
+    @model_validator(mode="after")
+    def validate_box_loader_inputs(self) -> Self:
         _box = None
 
         """Validate that we have either a box_developer_token or box_auth."""
-        if not values.get("box_auth") and not values.get("box_developer_token"):
+        if not self.box_auth and not self.box_developer_token:
             raise ValueError(
                 "you must provide box_developer_token or a box_auth "
                 "generated with langchain_box.utilities.BoxAuth"
             )
 
         _box = _BoxAPIWrapper(  # type: ignore[call-arg]
-            box_developer_token=values.get("box_developer_token"),
-            box_auth=values.get("box_auth"),
-            character_limit=values.get("character_limit"),
+            box_developer_token=self.box_developer_token,
+            box_auth=self.box_auth,
+            character_limit=self.character_limit,
         )
 
-        values["_box"] = _box
+        self._box = _box
 
-        return values
+        return self
 
     def _get_relevant_documents(
         self, query: str, *, run_manager: CallbackManagerForRetrieverRun
