@@ -1,15 +1,18 @@
 import logging
-from typing import Iterable, List, Optional
+from typing import Any, Iterable, List, Optional
 
 import voyageai  # type: ignore
 from langchain_core.embeddings import Embeddings
-from langchain_core.pydantic_v1 import (
-    BaseModel,
-    Field,
-    SecretStr,
-    root_validator,
-)
 from langchain_core.utils import secret_from_env
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    Field,
+    PrivateAttr,
+    SecretStr,
+    model_validator,
+)
+from typing_extensions import Self
 
 logger = logging.getLogger(__name__)
 
@@ -25,8 +28,8 @@ class VoyageAIEmbeddings(BaseModel, Embeddings):
             model = VoyageAIEmbeddings()
     """
 
-    _client: voyageai.Client = Field(exclude=True)
-    _aclient: voyageai.client_async.AsyncClient = Field(exclude=True)
+    _client: voyageai.Client = PrivateAttr()
+    _aclient: voyageai.client_async.AsyncClient = PrivateAttr()
     model: str
     batch_size: int
     show_progress_bar: bool = False
@@ -40,12 +43,14 @@ class VoyageAIEmbeddings(BaseModel, Embeddings):
         ),
     )
 
-    class Config:
-        extra = "forbid"
-        allow_population_by_field_name = True
+    model_config = ConfigDict(
+        extra="forbid",
+        populate_by_name=True,
+    )
 
-    @root_validator(pre=True)
-    def default_values(cls, values: dict) -> dict:
+    @model_validator(mode="before")
+    @classmethod
+    def default_values(cls, values: dict) -> Any:
         """Set default batch size based on model"""
         model = values.get("model")
         batch_size = values.get("batch_size")
@@ -53,13 +58,13 @@ class VoyageAIEmbeddings(BaseModel, Embeddings):
             values["batch_size"] = 72 if model in ["voyage-2", "voyage-02"] else 7
         return values
 
-    @root_validator(pre=False, skip_on_failure=True)
-    def validate_environment(cls, values: dict) -> dict:
+    @model_validator(mode="after")
+    def validate_environment(self) -> Self:
         """Validate that VoyageAI credentials exist in environment."""
-        api_key_str = values["voyage_api_key"].get_secret_value()
-        values["_client"] = voyageai.Client(api_key=api_key_str)
-        values["_aclient"] = voyageai.client_async.AsyncClient(api_key=api_key_str)
-        return values
+        api_key_str = self.voyage_api_key.get_secret_value()
+        self._client = voyageai.Client(api_key=api_key_str)
+        self._aclient = voyageai.client_async.AsyncClient(api_key=api_key_str)
+        return self
 
     def _get_batch_iterator(self, texts: List[str]) -> Iterable:
         if self.show_progress_bar:
