@@ -11,7 +11,6 @@ from langchain_core.prompts.prompt import PromptTemplate
 from langchain_core.runnables.base import Runnable, RunnableConfig
 from langchain_core.runnables.graph import Edge, Graph, Node
 from langchain_core.runnables.graph_mermaid import _escape_node_label
-from tests.unit_tests.pydantic_utils import _schema
 
 
 def test_graph_single_runnable(snapshot: SnapshotAssertion) -> None:
@@ -19,10 +18,10 @@ def test_graph_single_runnable(snapshot: SnapshotAssertion) -> None:
     graph = StrOutputParser().get_graph()
     first_node = graph.first_node()
     assert first_node is not None
-    assert _schema(first_node.data) == _schema(runnable.input_schema)  # type: ignore[union-attr]
+    assert first_node.data.schema() == runnable.get_input_jsonschema()  # type: ignore[union-attr]
     last_node = graph.last_node()
     assert last_node is not None
-    assert _schema(last_node.data) == _schema(runnable.output_schema)  # type: ignore[union-attr]
+    assert last_node.data.schema() == runnable.get_output_jsonschema()  # type: ignore[union-attr]
     assert len(graph.nodes) == 3
     assert len(graph.edges) == 2
     assert graph.edges[0].source == first_node.id
