@@ -2,7 +2,7 @@ from unittest.mock import Mock
 
 import pytest
 from langchain_core.documents import Document
-from pydantic.v1.error_wrappers import ValidationError
+from pydantic.error_wrappers import ValidationError
 from pytest_mock import MockerFixture
 
 from langchain_box.utilities import BoxAuth, BoxAuthType, _BoxAPIWrapper
