@@ -6,8 +6,8 @@ from langchain.agents.format_scratchpad import format_to_openai_function_message
 from langchain.tools import tool
 from langchain_core.language_models import FakeListLLM
 from langchain_core.messages import AIMessage, HumanMessage
-from langchain_core.pydantic_v1 import BaseModel, Field
 from langchain_core.utils.function_calling import convert_to_openai_function
+from pydantic import BaseModel, Field
 
 import langchain_prompty
 
